requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.5.0
newspaper3k>=0.2.8
serpapi>=0.1.5
scikit-learn>=1.4.0
numpy>=1.26.0
tqdm>=4.66.0
pinecone-client>=3.0.0
streamlit>=1.34.0
langchain
PyMuPDF
sentence-transformers
PyPDF2
tiktoken

lxml_html_clean
lxml[html_clean]
google-search-results

requests
beautifulsoup4
pdfplumber
python-magic
